Share pre-save argument delegation via a trait
UpdateModel and UpsertModel had byte-identical implementations.
