+++ src/Execution/Arguments/UpdateModel.php
@@ -9,6 +9,8 @@ use Nuwave\Lighthouse\Support\Contracts\PreSaveArgumentsAware;
 
 class UpdateModel implements ArgResolver, PreSaveArgumentsAware
 {
+    use DelegatesPreSaveArguments;
+
     public const MISSING_PRIMARY_KEY_FOR_UPDATE = 'Missing primary key for update.';
 
     /** @var callable|\Nuwave\Lighthouse\Support\Contracts\ArgResolver */
@@ -18,24 +20,6 @@ class UpdateModel implements ArgResolver, PreSaveArgumentsAware
     public function __construct(callable $previous)
     {
         $this->previous = $previous;
-    }
-
-    public function withPreSaveArguments(array $arguments): ?static
-    {
-        $previous = $this->previous;
-        if (! $previous instanceof PreSaveArgumentsAware) {
-            return null;
-        }
-
-        $previousWithPreSave = $previous->withPreSaveArguments($arguments);
-        if ($previousWithPreSave === null) {
-            return null;
-        }
-
-        $clone = clone $this;
-        $clone->previous = $previousWithPreSave;
-
-        return $clone;
     }
 
     /**
