+++ src/Execution/Arguments/UpsertModel.php
@@ -11,6 +11,8 @@ use function Safe\array_flip;
 
 class UpsertModel implements ArgResolver, PreSaveArgumentsAware
 {
+    use DelegatesPreSaveArguments;
+
     public const MISSING_IDENTIFYING_COLUMNS_FOR_UPSERT = 'All configured identifying columns must be present and non-null for upsert.';
 
     /** @var callable|\Nuwave\Lighthouse\Support\Contracts\ArgResolver */
@@ -23,24 +25,6 @@ class UpsertModel implements ArgResolver, PreSaveArgumentsAware
         protected ?array $identifyingColumns = null,
     ) {
         $this->previous = $previous;
-    }
-
-    public function withPreSaveArguments(array $arguments): ?static
-    {
-        $previous = $this->previous;
-        if (! $previous instanceof PreSaveArgumentsAware) {
-            return null;
-        }
-
-        $previousWithPreSave = $previous->withPreSaveArguments($arguments);
-        if ($previousWithPreSave === null) {
-            return null;
-        }
-
-        $clone = clone $this;
-        $clone->previous = $previousWithPreSave;
-
-        return $clone;
     }
 
     /**
